DAO,MangHinhLogin

+++ DTO/Passbook.cs
@@ -13,7 +13,7 @@ namespace DTO
         int _id;
         int _passbooktype;
         int _passbook_customer;
-        int _passbook_balance;
+        long _passbook_balance;
         DateTime? opendate;
 
         public Passbook() { }
@@ -22,7 +22,7 @@ namespace DTO
         {
             this.Id = int.Parse(row["ID"].ToString());
             this.Passbooktype = int.Parse(row["passbooktype"].ToString());
-            this.Passbook_balance = int.Parse(row["passbook_balance"].ToString());
+            this.Passbook_balance = long.Parse(row["passbook_balance"].ToString());
             this.Passbook_customer = int.Parse(row["passbook_customer"].ToString());
             this.Opendate = (DateTime?)row["opendate"];
 
@@ -32,7 +32,7 @@ namespace DTO
         public int Id { get => _id; set => _id = value; }
         public int Passbooktype { get => _passbooktype; set => _passbooktype = value; }
         public int Passbook_customer { get => _passbook_customer; set => _passbook_customer = value; }
-        public int Passbook_balance { get => _passbook_balance; set => _passbook_balance = value; }
+        public long Passbook_balance { get => _passbook_balance; set => _passbook_balance = value; }
 
         public DateTime? Opendate { get => opendate; set => opendate = value; }
     }
